Extract order rows
